# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GLCM Texture Analysis API",
    "description": "HTTP API of the GLCM texture analysis server (doc/ui-design-plan.md, section 8.5).",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "GLCM_API_TOKEN; required when the server sets one"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/images/{id}/roi-stats": {
      "post": {
        "summary": "Pixel count and intensity statistics of ROIs",
        "tags": [
          "rois"
        ],
        "description": "Masks follow the pixel-centre rule of glcm::RasterizeMask, so the counts equal those used by analyses.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "rois"
                ],
                "properties": {
                  "rois": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "id",
                        "shape"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "maxLength": 100
                        },
                        "shape": {
                          "anyOf": [
                            {
                              "type": "object",
                              "required": [
                                "type",
                                "x",
                                "y",
                                "width",
                                "height"
                              ],
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "enum": [
                                    "rectangle"
                                  ]
                                },
                                "x": {
                                  "type": "number",
                                  "description": "Image pixel coordinate; pixel (c, r) covers [c, c + 1) × [r, r + 1)"
                                },
                                "y": {
                                  "type": "number",
                                  "description": "Image pixel coordinate; pixel (c, r) covers [c, c + 1) × [r, r + 1)"
                                },
                                "width": {
                                  "type": "number"
                                },
                                "height": {
                                  "type": "number"
                                }
                              }
                            },
                            {
                              "type": "object",
                              "required": [
                                "type",
                                "cx",
                                "cy",
                                "rx",
                                "ry"
                              ],
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "enum": [
                                    "ellipse"
                                  ]
                                },
                                "cx": {
                                  "type": "number",
                                  "description": "Image pixel coordinate; pixel (c, r) covers [c, c + 1) × [r, r + 1)"
                                },
                                "cy": {
                                  "type": "number",
                                  "description": "Image pixel coordinate; pixel (c, r) covers [c, c + 1) × [r, r + 1)"
                                },
                                "rx": {
                                  "type": "number",
                                  "description": "Semi-axis along the ellipse x axis"
                                },
                                "ry": {
                                  "type": "number",
                                  "description": "Semi-axis along the ellipse y axis"
                                },
                                "angle": {
                                  "type": "number",
                                  "description": "Rotation in degrees, clockwise on screen"
                                }
                              }
                            },
                            {
                              "type": "object",
                              "required": [
                                "type",
                                "points"
                              ],
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "enum": [
                                    "polygon"
                                  ]
                                },
                                "points": {
                                  "type": "array",
                                  "items": {
                                    "type": "array",
                                    "additionalItems": false,
                                    "items": [
                                      {
                                        "type": "number",
                                        "description": "Image pixel coordinate; pixel (c, r) covers [c, c + 1) × [r, r + 1)"
                                      },
                                      {
                                        "type": "number",
                                        "description": "Image pixel coordinate; pixel (c, r) covers [c, c + 1) × [r, r + 1)"
                                      }
                                    ],
                                    "minItems": 2
                                  },
                                  "maxItems": 10000
                                },
                                "freehand": {
                                  "type": "boolean"
                                }
                              }
                            }
                          ]
                        }
                      }
                    },
                    "maxItems": 1000
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^img_[0-9a-f]{32}$"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Image id returned by POST /images"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "stats"
                  ],
                  "properties": {
                    "stats": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "roiId",
                          "pixelCount",
                          "boundingBox",
                          "min",
                          "max",
                          "mean",
                          "std",
                          "error"
                        ],
                        "properties": {
                          "roiId": {
                            "type": "string"
                          },
                          "pixelCount": {
                            "type": "integer",
                            "description": "Pixels whose centre lies inside the shape, clipped to the image"
                          },
                          "boundingBox": {
                            "anyOf": [
                              {
                                "type": "object",
                                "required": [
                                  "x",
                                  "y",
                                  "width",
                                  "height"
                                ],
                                "properties": {
                                  "x": {
                                    "type": "integer"
                                  },
                                  "y": {
                                    "type": "integer"
                                  },
                                  "width": {
                                    "type": "integer"
                                  },
                                  "height": {
                                    "type": "integer"
                                  }
                                }
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "min": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "max": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "mean": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "std": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "error": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ],
                            "description": "Invalid geometry"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error returned by every endpoint for 4xx and 5xx responses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable error code, e.g. \"NotFound\" or \"PayloadTooLarge\""
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable description"
                    }
                  },
                  "description": "Error returned by every endpoint for 4xx and 5xx responses"
                }
              }
            }
          },
          "404": {
            "description": "Error returned by every endpoint for 4xx and 5xx responses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable error code, e.g. \"NotFound\" or \"PayloadTooLarge\""
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable description"
                    }
                  },
                  "description": "Error returned by every endpoint for 4xx and 5xx responses"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analyses/{id}/events": {
      "get": {
        "summary": "Progress as Server-Sent Events",
        "tags": [
          "analyses"
        ],
        "description": "Events: \"result\" {index, result} for every finished job (earlier ones are replayed on connect), \"progress\" {completed, total}, and a final \"finished\" {status, completed, total, error}, after which the stream ends.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^ana_[0-9a-f]{32}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Event stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error returned by every endpoint for 4xx and 5xx responses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable error code, e.g. \"NotFound\" or \"PayloadTooLarge\""
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable description"
                    }
                  },
                  "description": "Error returned by every endpoint for 4xx and 5xx responses"
                }
              }
            }
          },
          "404": {
            "description": "Error returned by every endpoint for 4xx and 5xx responses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable error code, e.g. \"NotFound\" or \"PayloadTooLarge\""
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable description"
                    }
                  },
                  "description": "Error returned by every endpoint for 4xx and 5xx responses"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}